[LIR] [Docs] Peephole optimization

+++ src/lir/peephole/x86_64_gnu_nasm.c
@@ -18,6 +18,7 @@ static int _check_home(lir_block_t* h, lir_subject_t* s) {
     return 0;
 }
 
+// TODO: docs
 static int _jumps_pass(cfg_block_t* bb) {
     if (!bb->lmap.exit) return 0;
     lir_block_t* l  = LIR_get_back_instruction(bb->lmap.exit, bb->lmap.entry, 0);
@@ -36,6 +37,7 @@ static int _jumps_pass(cfg_block_t* bb) {
     return 1;
 }
 
+// TODO: docs
 static int _find_label_usage(cfg_func_t* fb, lir_subject_t* lb) {
     lir_block_t* lh = LIR_get_next(fb->lmap.entry, fb->lmap.exit, 0);
     while (lh) {
@@ -46,6 +48,7 @@ static int _find_label_usage(cfg_func_t* fb, lir_subject_t* lb) {
     return 0;
 }
 
+// TODO: docs
 static int _label_pass(cfg_func_t* fb) {
     lir_block_t* lh = LIR_get_next(fb->lmap.entry, fb->lmap.exit, 0);
     while (lh) {
@@ -56,7 +59,7 @@ static int _label_pass(cfg_func_t* fb) {
     return 1;
 }
 
-static unsigned int _visit_counter = 100;
+static unsigned long long _visit_counter = 100;
 
 /*
 Recursive cleanup will clean each block from the CFG with one simple rule:
@@ -154,6 +157,7 @@ int x86_64_gnu_nasm_peephole_optimization(cfg_ctx_t* cctx) {
             _jumps_pass(bb);
             _cleanup_pass(bb);
         }
+
         _label_pass(fb);
     }
